move templates to service
